+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2
+
+* Add "final", "CopyWith" and null handling.
+
+## 1.0.1
+
+* Starting the system for use - beta | testes.
+
 ## 1.0.0
 
 * Starting the system for use - beta.